[ADD] estate: added controllers and templates for website
To serve the users online we have created a website using odoo's controllers for
the backend, qweb templates, and Bootstrap to design an intuitive and easy-to-
use UI. Two controllers facilitate the viewing of all the properties and a
particular property respectively. The template for all the properties web page
has a grid layout where an overview can be taken of all the properties. Also
added the pager for a better user experience. Also added a new menu for the
website in the estate menus xml file. Added a static folder to store an image to
be showcased when there is no photo available. A date picker is added to the top
of property grid view which will be used to filter out the properties created
after the selected date by clicking on the listed after button