[ADD] estate: create model, security and UI
Learned about the folder structure of the module and created the estate
module in tutorials. Added name and description of the model and also added
varius type of fields in it. Create csv file for security purpose to restrict
access to data for the internal users. Make front end part for created model
using  XML file. learned basics about menuitem and record in it.
Completed 1-4 chapters in developer tutorials of Odoo.