[IMP] estate: define table relations, add menu and action
Added new tables in the estate such as property_type, property_tag
and property_offer.
Revised the concept of table creation and field creation.

Defines the relation between the tables such as Many2one, One2many and Many2many
Learned about relation and get know about how to make relation between the table
I have faced so many errors while writing the code for the relation between the
tables but with the reference of documents i have solved all of the errors.

Created new xml files to create the view created tables.
also added them in menu
Revised the concept of how to created views and menus

Added  actions for the model that newly added in estate module
Revised the concept of how to create action and add action to the model.