[IMP] estate: actions, action types, sql constraints and python constraints.
Created buttons like cancel and sold in the model and
added action to them and added some logic to them if offer is cancelled
than it should not be sold and vise versa.
Revised the concept of  actions and get know about the how action work
 with button
get to know about  property of the button like object type property
if we added to it button odoo framework will execute a python method
 link with action.
Added the icon in the list check mark for accept and cross mark for refuse
and also added business to them  logic if the it accepted the it can not refuse
and vise versa.  values
set the values of the selling price and set  the  buyer to the property
Again revised the concept of actions in buttons and written the business logic
 for set the values of the fields.
I took the reference of the examples in odoo docs and searched it on googles
and i also explored the code of odoo modules.
Added the sql constraint to the fields like check constraint and
unique constraint
Revised the concept of sql constraint and get know about how to add them using
python. Added python constraint to the field.
Learned the new concept of python constraint.