[IMP] estate: sprinkles-  search and stat button inheritance.
Completed 11 chapter of sprinkles added default filter and filter domain
 in the search bar.
Learned how to add default filter and  filter domain so user can search by
property.

Added stat button in the property type form when we click on it which
shows the list of all related record.
It was challenging because we have ti define the action for that
 button and map them to the referring model and also search for the class to
 added the button.
started  chapter 12 inheritance written a method so it could not let delete
the record to user when it in particular state
learned about the inheritance.