[IMP] estate, estate_account: link module, qweb , final word and demo data
Completed 13 chapter  created invoice and added 2 line in it.
It was challenging because I have to inherent the model and
override the sold method I took the reference of the odoo code base and final
I done it.
I Learned how to create invoice.

Also added two lines in the invoice by defining the relation in between both
the inherited model.
it was also bit challenging  and I leaned how to add lines in the invoice.

Completed 14 chapter  Created Kanban view in estate module , improve the
Kanban view and add the default grouping
I done it by taking the reference of odoo code base and I learn how to create
kanban view.

Finally done the chapter 15 make the changes according to the coding guidelines

Started adding demo data in the module.
Added csv file and xml files using them i added data in my module
it was also challenging and at last i done it.