[IMP] estate: added access rights and report generation for properties
A user can now generate reports of their properties where they can
easily view the offers made to the property as well along with other info
like the status, best price etc.
This generated PDF report can be sent to the customers as well as upper
management.
A user can also generate a report by going to a specific user and see all the
properties where the user is listed as a salesman.
Also implemented Strict ACLs based on the fact that the user is
an agent or a manager.