[IMP] estate: menu  in website, properties detail and pagers and warranty config
Added properties menu in website that will display available properties
so the customer can check the available properties on website.
Learned how to add menu on the website and display the available properties.

On click  on the view details button that will display the all the details of
the available properties
Learned how to route the pages.

Added pagers in the properties view so with help of it user can go on next page
and can come back to previous
Learned how to add the pagers.

started  new  warranty configuration model added menu  in sale module.