[IMP] warranty: add warranty line in order line and delete warranty line
Added method in add wizard model that will add line in order line when on click
add button
it was challenging because i have to create line and refer the correct id that
have relation between the models to create  line at end i done it
learn how yo add line in model from transient model.

Added on delete cascade field in order line that will have the self join
it will delete child when parent deleted.
learn how to implement the on delete cascade and how to make self
join  in model.