[IMP] budget: added new views and upgraded logic
A user can now create multiple budgets on a quarterly or monthly basis
with any no. of selected accounts and define a budget amount for them.
What this will do is whenever a new expense occurs from their account,
the achieved amount and percentage will automatically get calculated which
will either give them a warning or error that they are over budget.
This behavior is configurable too.
On top of it, they can visualize the budget lines/ analytical accounts on a no. of
metrics through a list, pivot or graph view.