[IMP] awesome_clicker: create a client action and move the state to a service.
The client action will display a dedicated window that will serve to play the game. In the current state, it only contains an increment button with a total value of 10.
The clicks state has been transfer to a service to share it between the systray item and the client action. That way the clicks count will be the same for the entire web client.
The service also implement the increment() function to replace the one that was defined in the systray item class.