[IMP] awesome_clicker: add a custom clicker hook, humanize the clicks counter and add the ClickBots.
This will introduce a custom cliker hook that will handle the required import to create and use the clicker service. That way, the classes that use this service will only require to import and use this hook.
This will also add the clickValue class and component. It transform the displayed value of the clicks counter to something more readable for human, especially with greater values. To allow the player to see the exact number of clicks, a dynamic tooltip has been added to the counter.
Finally, this will add the possibility to buy ClickBots with 1000 clicks. Those ClickBots will be available when reaching level 1 (1000 clicks) and will generate 10 clicks each every 10 seconds.