[IMP] awesome_clicker: add a new power resource.
This will add a new resource to the game: power upgrades.
This upgrade will unlock at level 3 (100k clicks) and will cost 50k for each. It multiplies the power of the bots.
Thi will also add the total clicks per second of every bot and the grand total on the client action view.