[IMP] awesome_clicker: add tree resource and commands.
This will add a new resource: trees. Each tree will cost 1M clicks and produce a fruit each 30 seconds. It represent the actual end-game of the clicker.
The systray item has been improve to display the click count, tree count and fruit count and transform it into a dropdown menu.
Inside the dropdown, we can access the clicker game panel and use a command to buy a clickbot. It also display the number of purchased trees and produced fruits in more detail.